renamed bondManager -> substrateCallManager
